Move the search controller to the api namespace
